prefetchRegistrationAndEvent helper function to prefetch registration and event data
