+++ src/pages/registration/[registrationId]/enrolment/create/index.tsx
@@ -3,12 +3,8 @@ import { GetServerSideProps, NextPage } from 'next';
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
 
 import CreateEnrolmentPage from '../../../../../domain/enrolment/CreateEnrolmentPage';
-import { EVENT_INCLUDES } from '../../../../../domain/event/constants';
-import { fetchEventQuery } from '../../../../../domain/event/query';
-import { prefetchPlaceQuery } from '../../../../../domain/place/query';
-import { fetchRegistrationQuery } from '../../../../../domain/registration/query';
 import { getSessionAndUser } from '../../../../../utils/getSessionAndUser';
-import parseIdFromAtId from '../../../../../utils/parseIdFromAtId';
+import prefetchRegistrationAndEvent from '../../../../../utils/prefetchRegistrationAndEvent';
 
 const CreateEnrolment: NextPage = () => <CreateEnrolmentPage />;
 
@@ -24,29 +20,13 @@ export const getServerSideProps: GetServerSideProps = async ({
     res,
   });
 
-  try {
-    const registration = await fetchRegistrationQuery(
-      queryClient,
-      { id: query.registrationId as string },
-      { req, res }
-    );
-
-    if (registration?.event) {
-      const event = await fetchEventQuery(
-        queryClient,
-        { id: registration?.event, include: EVENT_INCLUDES },
-        { req, res }
-      );
-
-      if (event) {
-        const placeId = parseIdFromAtId(event.location['@id']);
-
-        if (placeId) {
-          await prefetchPlaceQuery(queryClient, placeId, { req, res });
-        }
-      }
-    }
-  } catch {}
+  await prefetchRegistrationAndEvent({
+    query,
+    queryClient,
+    req,
+    res,
+    shouldPrefetchPlace: true,
+  });
 
   return {
     props: {
